+++ pkg/controller/scylladbcluster/resource_test.go
@@ -565,6 +565,7 @@ func TestMakeScyllaDBDatacenters(t *testing.T) {
 						Name: "c",
 					},
 				},
+				BootstrapPolicy: pointer.Ptr(scyllav1alpha1.BootstrapPolicySequential),
 			},
 		}
 	}